requirements.txt
langchain_chroma==0.1.2
langchain_community==0.2.11
langchain_openai==0.1.21
langchain_text_splitters==0.2.2
langchain_core==0.2.29
pypdf==4.3.1
cryptography==43.0.0
python-dotenv==1.0.1
sudachipy==0.6.8
sudachidict_core
rank_bm25==0.2.2
langgraph==0.2.16
pdfplumber==0.11.4
